refactor(asset): 移动 asset-loader.ts 和 asset-manager.ts 文件至主目录
