+++ packages/effects-core/src/asset/index.ts
@@ -1,3 +1,1 @@
-export * from './asset-loader';
-export * from './asset-manager';
 export * from './asset';
